Split station_test.exs to station_contract_test.exs, station_processing_test.exs, station_correctness_test.exs and station_stress_test.exs
