+++ .travis.yml
@@ -2,7 +2,7 @@ language: elixir
 elixir:
 - 1.6.2
 otp_release:
-- 20.0
+- 20.3
 
 install:
 - bash script/travis-install
