+++ test/test_helper.exs
@@ -1,1 +1,4 @@
 ExUnit.start()
+Mox.defmock(MockCollector, for: Station.Collector)
+Mox.defmock(MockRegister, for: Station.Registry)
+Mox.defmock(MockStation, for: Station.StationBehaviour)
